cpa/requirements.txt
addict==2.4.0
albumentations==1.4.1
boto3==1.34.62
botocore==1.34.62
cityscapesscripts==2.2.2
clip==0.2.0
einops==0.7.0
fairscale==0.4.13
ffmpegcv==0.3.11
gradio==4.21.0
imagecorruptions==1.1.2
imageio==2.31.4
instaboostfast==0.1.2
label_studio_ml==1.0.9
label_studio_tools==0.0.3
lap==0.4.0
matplotlib==3.7.4
memory_profiler==0.61.0
mmcv==2.1.0
mmengine==0.10.1
mmpretrain==1.2.0
model_archiver==1.0.3
model_index==0.1.11
motmetrics==1.4.0
nltk==3.8.1
numpy==1.24.3
opencv_python==4.8.1.78
openpyxl==3.1.2
pandas==2.0.3
parameterized==0.9.0
Pillow==10.0.1
prettytable==3.10.0
psutil==5.9.0
pycocoevalcap==1.2
pycocotools==2.0.7
pytest==8.1.1
pytorch_sphinx_theme==0.0.19
PyYAML==6.0.1
Requests==2.31.0
rich==13.7.1
roboflow==1.1.23
sahi==0.11.15
scikit_image==0.19.3
scikit_learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==60.2.0
Shapely==2.0.3
six==1.16.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.2
transformers==4.38.2
ts==0.5.1
wandb==0.16.1
xlrd==2.0.1
xlutils==2.0.0
